updated_UI
Added date,total,quantity, merchant fields that are filled with info from the receipt. The regex should be reviewed.
Also made a bottom navbar for switching between views, I intend to add in the receipts view the date and then the shop with the total, then you could click for more details.  I have to implement a database for storing the items first